# service/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinic Antiviral Reporter",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/sessions": {
      "post": {
        "summary": "Create Session",
        "operationId": "createSession",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSession"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SessionView"
                    },
                    {
                      "$ref": "#/components/schemas/PairingView"
                    },
                    {
                      "$ref": "#/components/schemas/DeviceView"
                    },
                    {
                      "$ref": "#/components/schemas/RefreshView"
                    },
                    {
                      "$ref": "#/components/schemas/CaseView"
                    },
                    {
                      "$ref": "#/components/schemas/BulkLotView"
                    },
                    {
                      "$ref": "#/components/schemas/ScanView"
                    },
                    {
                      "$ref": "#/components/schemas/MappingView"
                    }
                  ],
                  "title": "Response Createsession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/session": {
      "get": {
        "summary": "Get Session",
        "operationId": "getSession",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices": {
      "get": {
        "summary": "List Devices",
        "operationId": "listDevices",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceView"
                  },
                  "title": "Response Listdevices"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "summary": "List Audit",
        "operationId": "listAuditEvents",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditView"
                  },
                  "title": "Response Listauditevents"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pairings": {
      "post": {
        "summary": "Create Pairing",
        "operationId": "createPairing",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePairing"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SessionView"
                    },
                    {
                      "$ref": "#/components/schemas/PairingView"
                    },
                    {
                      "$ref": "#/components/schemas/DeviceView"
                    },
                    {
                      "$ref": "#/components/schemas/RefreshView"
                    },
                    {
                      "$ref": "#/components/schemas/CaseView"
                    },
                    {
                      "$ref": "#/components/schemas/BulkLotView"
                    },
                    {
                      "$ref": "#/components/schemas/ScanView"
                    },
                    {
                      "$ref": "#/components/schemas/MappingView"
                    }
                  ],
                  "title": "Response Createpairing"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/enroll": {
      "post": {
        "summary": "Enroll Device",
        "operationId": "enrollDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollDevice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}/revoke": {
      "post": {
        "summary": "Revoke Device",
        "operationId": "revokeDevice",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeDevice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SessionView"
                    },
                    {
                      "$ref": "#/components/schemas/PairingView"
                    },
                    {
                      "$ref": "#/components/schemas/DeviceView"
                    },
                    {
                      "$ref": "#/components/schemas/RefreshView"
                    },
                    {
                      "$ref": "#/components/schemas/CaseView"
                    },
                    {
                      "$ref": "#/components/schemas/BulkLotView"
                    },
                    {
                      "$ref": "#/components/schemas/ScanView"
                    },
                    {
                      "$ref": "#/components/schemas/MappingView"
                    }
                  ],
                  "title": "Response Revokedevice"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkLotView": {
        "properties": {
          "appliedCases": {
            "items": {
              "$ref": "#/components/schemas/CaseView"
            },
            "type": "array",
            "title": "Appliedcases"
          }
        },
        "type": "object",
        "required": [
          "appliedCases"
        ],
        "title": "BulkLotView"
      },
...
      "CreateSession": {
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid",
            "title": "Requestid"
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expectedrevision"
          },
          "operator": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Operator"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "expectedRevision",
          "operator"
        ],
        "title": "CreateSession"
      },
      "DeviceView": {
        "properties": {
          "deviceId": {
            "type": "string",
            "title": "Deviceid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "physician",
                "reporting"
              ]
            },
            "type": "array",
            "title": "Capabilities"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "revoked": {
            "type": "boolean",
            "title": "Revoked"
          }
        },
        "type": "object",
        "required": [
          "deviceId",
          "name",
          "capabilities",
          "revision",
          "revoked"
        ],
        "title": "DeviceView"
      },
...
      "PairingView": {
        "properties": {
          "pairingCode": {
            "type": "string",
            "title": "Pairingcode"
          },
          "expiresAt": {
            "type": "number",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "pairingCode",
          "expiresAt"
        ],
        "title": "PairingView"
      },
...
      "RefreshView": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "unchanged": {
            "type": "integer",
            "title": "Unchanged"
          },
          "changed": {
            "type": "integer",
            "title": "Changed",
            "default": 0
          },
          "quarantined": {
            "type": "integer",
            "title": "Quarantined",
            "default": 0
          },
          "synthetic": {
            "type": "boolean",
            "const": true,
            "title": "Synthetic",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "revision",
          "created",
          "unchanged"
        ],
        "title": "RefreshView"
      },
...
      "ScanView": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "status": {
            "type": "string",
            "enum": [
              "disabled",
              "awaiting_configuration",
              "idle",
              "queued",
              "running",
              "succeeded",
              "partial",
              "failed",
              "interrupted"
            ],
            "title": "Status"
          },
          "jobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobid"
          },
          "dateFrom": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datefrom"
          },
          "dateTo": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateto"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "lastSuccessAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastsuccessat"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": true
          },
          "diagnostic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostic"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts",
            "default": {}
          },
          "intervalSeconds": {
            "type": "integer",
            "title": "Intervalseconds"
          },
          "synthetic": {
            "type": "boolean",
            "const": true,
            "title": "Synthetic",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "revision",
          "enabled",
          "status",
          "intervalSeconds"
        ],
        "title": "ScanView"
      },
      "SessionView": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "deviceId": {
            "type": "string",
            "title": "Deviceid"
          },
          "operator": {
            "type": "string",
            "title": "Operator"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "expiresAt": {
            "type": "number",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "deviceId",
          "operator",
          "capabilities",
          "revision",
          "expiresAt"
        ],
        "title": "SessionView"
      },
...
    }
  }
}